Refactor : CacheService 생성및 캐시 로직 따로 분리
