+++ src/main/java/com/giftforyoube/funding/service/FundingService.java
@@ -1,8 +1,6 @@
 package com.giftforyoube.funding.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.giftforyoube.donation.repository.DonationRepository;
 import com.giftforyoube.funding.dto.*;
 import com.giftforyoube.funding.entity.Funding;
@@ -23,20 +21,13 @@ import org.jsoup.nodes.Document;
 import org.jsoup.select.Elements;
 import org.redisson.api.RLock;
 import org.redisson.api.RedissonClient;
-import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.domain.*;
-import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.io.IOException;
-import java.time.Duration;
 import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.Set;
 import java.util.concurrent.TimeUnit;
 
 @Slf4j
@@ -45,17 +36,14 @@ import java.util.concurrent.TimeUnit;
 @RequiredArgsConstructor
 public class FundingService {
 
-    private final RedisTemplate<String, String> redisTemplate;
     private final FundingRepository fundingRepository;
-    private final ObjectMapper objectMapper;
     private final UserRepository userRepository;
     private final RedissonClient redissonClient;
     private final DonationRepository donationRepository;
     private final FundingSummaryRepository fundingSummaryRepository;
+    private final CacheService cacheService;
 
     private static final int TIMEOUT = 10000; // 10초
-    private static final String FUNDING_ITEM_CACHE_PREFIX = "cachedFundingItem:";
-    private static final String FUNDING_SUMMARY_CACHE_KEY = "fundingSummary";
 
     // 데이터베이스 트랜잭션에 직접적으로 관련된 작업이 없으므로 @Transactional 어노테이션을 사용할 필요가 없음.
     public FundingItemResponseDto addLinkAndSaveToCache(AddLinkRequestDto requestDto, Long userId) throws IOException {
@@ -70,7 +58,7 @@ public class FundingService {
                 throw new IllegalStateException("해당 사용자에 대한 락을 획득할 수 없습니다 : " + userId);
             }
             FundingItem fundingItem = previewItem(requestDto.getItemLink());
-            saveToCache(fundingItem, userId.toString());
+            cacheService.saveToCache(fundingItem, userId.toString());
             return FundingItemResponseDto.fromEntity(fundingItem);
         } catch (InterruptedException e) {
             Thread.currentThread().interrupt();
@@ -101,8 +89,8 @@ public class FundingService {
             if (hasActiveFunding) {
                 throw new IllegalStateException("이미 진행중인 펀딩이 있습니다.");
             }
-            String userCacheKey = buildCacheKey(userId.toString());
-            FundingItem fundingItem = getCachedFundingProduct(userCacheKey);
+            String userCacheKey = cacheService.buildCacheKey(userId.toString());
+            FundingItem fundingItem = cacheService.getCachedFundingProduct(userCacheKey);
             if (fundingItem == null) {
                 throw new IllegalStateException("링크 상품을 찾을 수 없습니다.");
             }
@@ -111,8 +99,8 @@ public class FundingService {
             Funding funding = requestDto.toEntity(fundingItem, status);
             funding.setUser(user);
             fundingRepository.save(funding);
-            clearCache(userCacheKey);
-            clearFundingCaches();
+            cacheService.clearCache(userCacheKey);
+            cacheService.clearFundingCaches();
             return FundingResponseDto.fromEntity(funding);
         } catch (InterruptedException e) {
             Thread.currentThread().interrupt();
@@ -128,7 +116,7 @@ public class FundingService {
     public FundingResponseDto findFunding(Long fundingId) {
         String cacheKey = "fundingDetail:" + fundingId;
         // 캐시에서 조회 시도
-        FundingResponseDto cachedFunding = getFundingFromCache(cacheKey);
+        FundingResponseDto cachedFunding = cacheService.getFundingFromCache(cacheKey);
         if (cachedFunding != null) {
             return cachedFunding;
         }
@@ -139,7 +127,7 @@ public class FundingService {
         FundingResponseDto fundingResponseDto = FundingResponseDto.fromEntity(funding);
 
         // 결과를 캐시에 저장
-        saveFundingToCache(cacheKey, fundingResponseDto);
+        cacheService.saveFundingToCache(cacheKey, fundingResponseDto);
         return fundingResponseDto;
     }
 
@@ -150,7 +138,7 @@ public class FundingService {
 
         String cacheKey = "fundingDetail:" + currentUser.getId();
         // 캐시에서 조회 시도
-        FundingResponseDto cachedFunding = getFundingFromCache(cacheKey);
+        FundingResponseDto cachedFunding = cacheService.getFundingFromCache(cacheKey);
         if (cachedFunding != null) {
             return cachedFunding;
         }
@@ -162,7 +150,7 @@ public class FundingService {
 
         FundingResponseDto fundingResponseDto = FundingResponseDto.fromEntity(funding);
         // 결과를 캐시에 저장
-        saveFundingToCache(cacheKey, fundingResponseDto);
+        cacheService.saveFundingToCache(cacheKey, fundingResponseDto);
 
         return fundingResponseDto;
     }
@@ -175,7 +163,7 @@ public class FundingService {
         String cacheKey = "activeMainFundings:" + page + ":" + size + ":" + sortBy + ":" + sortOrder;
 
         // 캐시에서 조회 시도
-        Page<FundingResponseDto> cachedPage = getFundingPageFromCache(cacheKey, pageable);
+        Page<FundingResponseDto> cachedPage = cacheService.getFundingPageFromCache(cacheKey, pageable);
         if (cachedPage != null) {
             return cachedPage;
         }
@@ -185,7 +173,7 @@ public class FundingService {
         Page<FundingResponseDto> fundingResponseDtoPage = mainFundings.map(FundingResponseDto::fromEntity);
 
         // 결과를 캐시에 저장
-        saveFundingPageToCache(cacheKey, fundingResponseDtoPage);
+        cacheService.saveFundingPageToCache(cacheKey, fundingResponseDtoPage);
 
         return fundingResponseDtoPage;
     }
@@ -196,7 +184,7 @@ public class FundingService {
         String cacheKey = "allFundings:" + page + ":" + size + ":" + sortBy + ":" + sortOrder;
 
         // 캐시에서 조회 시도
-        Page<FundingResponseDto> cachedFundings = getFundingsPageFromCache(cacheKey, pageable);
+        Page<FundingResponseDto> cachedFundings = cacheService.getFundingsPageFromCache(cacheKey, pageable);
         if (cachedFundings != null && !cachedFundings.isEmpty()) {
             return cachedFundings;
         }
@@ -206,7 +194,7 @@ public class FundingService {
         Page<FundingResponseDto> allFundings = allFunding.map(FundingResponseDto::fromEntity);
 
         // 결과를 캐시에 저장
-        saveFundingsPageToCache(cacheKey, allFundings);
+        cacheService.saveFundingsPageToCache(cacheKey, allFundings);
 
         return allFundings;
     }
@@ -218,14 +206,14 @@ public class FundingService {
         String cacheKey = "activeFundings:" + page + ":" + size + ":" + sortBy + ":" + sortOrder;
 
         // 캐시에서 조회 시도
-        Slice<FundingResponseDto> cachedFundings = getFundingListFromCache(cacheKey, pageable);
+        Slice<FundingResponseDto> cachedFundings = cacheService.getFundingListFromCache(cacheKey, pageable);
         if (cachedFundings != null && !cachedFundings.isEmpty()) {
             return cachedFundings;
         }
 
         // DB에서 조회 및 캐시 저장
         Slice<FundingResponseDto> activeFundings = fundingRepository.findByStatus(FundingStatus.ACTIVE, pageable).map(FundingResponseDto::fromEntity);
-        saveFundingListToCache(cacheKey, activeFundings);
+        cacheService.saveFundingListToCache(cacheKey, activeFundings);
 
         return activeFundings;
     }
@@ -238,14 +226,14 @@ public class FundingService {
         String cacheKey = "finishedFundings:" + page + ":" + size + ":" + sortBy + ":" + sortOrder;
 
         // 캐시에서 조회 시도
-        Slice<FundingResponseDto> cachedFundings = getFundingListFromCache(cacheKey, pageable);
+        Slice<FundingResponseDto> cachedFundings = cacheService.getFundingListFromCache(cacheKey, pageable);
         if (!cachedFundings.getContent().isEmpty()) {
             return cachedFundings;
         }
 
         // DB에서 조회 및 캐시 저장
         Slice<FundingResponseDto> finishedFundings = fundingRepository.findByStatus(FundingStatus.FINISHED, pageable).map(FundingResponseDto::fromEntity);
-        saveFundingListToCache(cacheKey, finishedFundings);
+        cacheService.saveFundingListToCache(cacheKey, finishedFundings);
 
         return finishedFundings;
     }
@@ -264,7 +252,7 @@ public class FundingService {
 
         funding.setStatus(FundingStatus.FINISHED);
         fundingRepository.save(funding);
-        clearFundingCaches();
+        cacheService.clearFundingCaches();
     }
 
     // 펀딩 수정
@@ -289,13 +277,15 @@ public class FundingService {
             }
 
             funding.update(requestDto); // 펀딩 내용수정
-            clearFundingCaches(); // 캐시 무효화
+            cacheService.clearFundingCaches(); // 캐시 무효화
             return FundingResponseDto.fromEntity(funding);
         } catch (InterruptedException e) {
             Thread.currentThread().interrupt();
             throw new IllegalStateException("락을 획득하는 동안 문제가 발생하였습니다.", e);
         } finally {
-            lock.unlock();
+            if(lock.isLocked() && lock.isHeldByCurrentThread()){
+                lock.unlock();
+            }
         }
     }
 
@@ -322,12 +312,14 @@ public class FundingService {
             }
 
             fundingRepository.delete(funding);
-            clearFundingCaches(); // 캐시 무효화
+            cacheService.clearFundingCaches(); // 캐시 무효화
         } catch (InterruptedException e) {
             Thread.currentThread().interrupt();
             throw new IllegalStateException("락을 획득하는 동안 문제가 발생하였습니다.", e);
         } finally {
-            lock.unlock();
+            if(lock.isLocked() && lock.isHeldByCurrentThread()){
+                lock.unlock();
+            }
         }
     }
 
@@ -335,7 +327,7 @@ public class FundingService {
     @Transactional(readOnly = true)
     public FundingSummaryResponseDto getFundingSummary() {
         // 캐시에서 통계 데이터를 검색합니다.
-        FundingSummaryResponseDto cachedSummary = getSummaryFromCache(FUNDING_SUMMARY_CACHE_KEY);
+        FundingSummaryResponseDto cachedSummary = cacheService.getSummaryFromCache();
         if (cachedSummary != null) {
             return cachedSummary;
         }
@@ -353,36 +345,12 @@ public class FundingService {
                 .totalFundingAmount(totalFundingAmount)
                 .build();
 
-        saveSummaryToCache(FUNDING_SUMMARY_CACHE_KEY, summary);
+        cacheService.saveSummaryToCache(summary);
         return summary;
     }
 
-    // 목표금액 달성되어서 종료된 펀딩 카운트 증가 메서드 추가 예정
+    // ---------------------------- OG 태그 메서드 ------------------------------------------
 
-
-
-    // ---------------------------- 캐시 관련 메서드들과 OG 태그 메서드 ------------------------------------------
-
-    private String buildCacheKey(String userId) {
-        return FUNDING_ITEM_CACHE_PREFIX + userId;
-    }
-
-    // FundingItem 객체를 JSON으로 변환하여 캐시에 저장
-    public void saveToCache(FundingItem fundingItem, String userId) throws JsonProcessingException {
-        log.info("[saveToCache] 캐쉬에 저장하기");
-
-        String cacheKey = buildCacheKey(userId);
-        String fundingItemJson = objectMapper.writeValueAsString(fundingItem);
-        redisTemplate.opsForValue().set(cacheKey, fundingItemJson, Duration.ofDays(1));
-    }
-
-    // 캐시에서 FundingItem 객체를 가져오기
-    public FundingItem getCachedFundingProduct(String cacheKey) throws JsonProcessingException {
-        log.info("[getCachedFundingProduct] 캐시에서 FundingItem 객체를 가져오기");
-
-        String fundingItemJson = redisTemplate.opsForValue().get(cacheKey); // 역직렬화
-        return fundingItemJson == null ? null : objectMapper.readValue(fundingItemJson, FundingItem.class);
-    }
 
     public FundingItem previewItem(String itemLink) throws IOException {
         log.info("[previewItem] 상품 미리보기");
@@ -395,13 +363,6 @@ public class FundingService {
         return new FundingItem(itemLink, itemImage);
     }
 
-    public void clearCache(String userCacheKey) {
-        log.info("[clearCache] 캐쉬 삭제하기");
-
-        String cacheKey = buildCacheKey(userCacheKey);
-        redisTemplate.delete(cacheKey);
-    }
-
     private static String getMetaTagContent(Document document, String property) {
         log.info("[getMetaTagContent] 메타 태크에서 상품이미지 가져오기");
 
@@ -410,164 +371,5 @@ public class FundingService {
             return metaTags.first().attr("content");
         }
         return null;
-    }
-
-
-    //    // 캐시에 Page 데이터 저장
-    private void saveFundingPageToCache(String cacheKey, Page<FundingResponseDto> page) {
-        try {
-            // Page 구현체를 JSON으로 변환하는 과정에서는 구현체의 구체적인 클래스 정보가 필요할 수 있으므로,
-            // Page 내용만 캐시하고, 페이징 정보는 별도로 관리하는 것을 고려해야 할 수 있습니다.
-            String jsonContent = objectMapper.writeValueAsString(page.getContent());
-            redisTemplate.opsForValue().set(cacheKey, jsonContent, Duration.ofHours(1));
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException("Error serializing funding page data", e);
-        }
-    }
-
-    // 캐시에서 Page 데이터 조회
-    private Page<FundingResponseDto> getFundingPageFromCache(String cacheKey, Pageable pageable) {
-        String jsonContent = redisTemplate.opsForValue().get(cacheKey);
-        if (jsonContent == null) {
-            return null;
-        }
-        try {
-            List<FundingResponseDto> content = objectMapper.readValue(jsonContent, new TypeReference<List<FundingResponseDto>>(){});
-            // 여기서는 캐시된 내용과 Pageable 정보를 기반으로 새 Page 객체를 생성해야 합니다.
-            // 실제 페이지 크기와 전체 페이지 수 등은 DB 조회 없이 알 수 없으므로, 이 부분은 적절히 조정이 필요합니다.
-            return new PageImpl<>(content, pageable, content.size());
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException("Error deserializing funding page data", e);
-        }
-    }
-    // 캐시 관련 메서드 수정
-    public void saveFundingsPageToCache(String cacheKey, Page<FundingResponseDto> page) {
-        try {
-            FundingPageCached<FundingResponseDto> cachedPage = new FundingPageCached<>();
-            cachedPage.setContent(page.getContent());
-            cachedPage.setMetadata(new FundingPageMetadata(page.getTotalPages(), page.getTotalElements()));
-
-            String jsonContent = objectMapper.writeValueAsString(cachedPage);
-            redisTemplate.opsForValue().set(cacheKey, jsonContent, Duration.ofHours(1));
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException("Error serializing funding page data", e);
-        }
-    }
-
-    public Page<FundingResponseDto> getFundingsPageFromCache(String cacheKey, Pageable pageable) {
-        String jsonContent = redisTemplate.opsForValue().get(cacheKey);
-        if (jsonContent == null) {
-            return Page.empty(); // 캐시에서 데이터를 가져올 수 없으면 빈 페이지 반환
-        }
-        try {
-            FundingPageCached<FundingResponseDto> cachedPage = objectMapper.readValue(jsonContent, new TypeReference<FundingPageCached<FundingResponseDto>>() {
-            });
-            return new PageImpl<>(cachedPage.getContent(), pageable, cachedPage.getMetadata().getTotalElements());
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException("Error deserializing funding page data", e);
-        }
-    }
-
-
-    // 캐시에 펀딩 목록 저장하는 로직
-    public void saveFundingListToCache(String cacheKey, Slice<FundingResponseDto> fundings) {
-        FundingResponseDtoCache cache = new FundingResponseDtoCache(
-                new ArrayList<>(fundings.getContent()),
-                fundings.getNumber(),
-                fundings.getSize(),
-                fundings.isLast()
-        );
-
-        try {
-            String jsonContent = objectMapper.writeValueAsString(cache);
-            redisTemplate.opsForValue().set(cacheKey, jsonContent, Duration.ofHours(1)); // 캐시 만료 시간은 필요에 따라 조정
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException("Error serializing funding data", e);
-        }
-    }
-
-    // 캐시에서 목록 조회하는 로직
-    public Slice<FundingResponseDto> getFundingListFromCache(String cacheKey, Pageable pageable) {
-        String jsonContent = redisTemplate.opsForValue().get(cacheKey);
-        if (jsonContent == null) {
-            return new SliceImpl<>(Collections.emptyList(), pageable, false);
-        }
-        try {
-            FundingResponseDtoCache cache = objectMapper.readValue(jsonContent, FundingResponseDtoCache.class);
-            return new SliceImpl<>(cache.getContent(), PageRequest.of(cache.getPage(), cache.getSize()), cache.isLast());
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException("Error deserializing funding data", e);
-        }
-    }
-
-    // 펀딩 상세 정보 캐시에 저장
-    private void saveFundingToCache(String cacheKey, FundingResponseDto fundingResponseDto) {
-        try {
-            String jsonContent = objectMapper.writeValueAsString(fundingResponseDto);
-            redisTemplate.opsForValue().set(cacheKey, jsonContent, Duration.ofHours(1));
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException("Error serializing funding detail", e);
-        }
-    }
-
-    // 펀딩 상세 정보 캐시에서 조회
-    private FundingResponseDto getFundingFromCache(String cacheKey) {
-        String jsonContent = redisTemplate.opsForValue().get(cacheKey);
-        if (jsonContent == null) {
-            return null;
-        }
-        try {
-            return objectMapper.readValue(jsonContent, FundingResponseDto.class);
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException("Error deserializing funding detail", e);
-        }
-    }
-
-    // Giftipie에서 함께한 선물 캐시에서 가져오기
-    private FundingSummaryResponseDto getSummaryFromCache(String cacheKey) {
-        String jsonContent = redisTemplate.opsForValue().get(cacheKey);
-        if (jsonContent == null) {
-            return null;
-        }
-        try {
-            return objectMapper.readValue(jsonContent, FundingSummaryResponseDto.class);
-        } catch (JsonProcessingException e) {
-            log.error("Error deserializing funding summary from cache", e);
-            return null;
-        }
-    }
-
-    // Giftipie에서 함께한 선물 캐시에 저장
-    private void saveSummaryToCache(String cacheKey, FundingSummaryResponseDto summary) {
-        try {
-            String jsonContent = objectMapper.writeValueAsString(summary);
-            redisTemplate.opsForValue().set(cacheKey, jsonContent, Duration.ofHours(1)); // 캐시 유지 시간은 요구 사항에 따라 조정 가능
-        } catch (JsonProcessingException e) {
-            log.error("Error serializing funding summary to cache", e);
-        }
-    }
-
-    // 펀딩 생성, 업데이트, 삭제 시 캐시 삭제
-    public void clearFundingCaches() {
-        // 메인 펀딩 관련 캐시 삭제
-        clearCacheByPattern("activeMainFundings:*");
-
-        // 기존의 펀딩 리스트 관련 캐시 삭제
-        clearCacheByPattern("allFundings:*");
-        clearCacheByPattern("activeFundings:*");
-        clearCacheByPattern("finishedFundings:*");
-
-        // 상세 페이지 캐시 삭제 추가
-        clearCacheByPattern("fundingDetail:*");
-
-        // Giftipie에서 함께한 선물 캐시 삭제
-        clearCacheByPattern(FUNDING_SUMMARY_CACHE_KEY);
-    }
-
-    private void clearCacheByPattern(String pattern) {
-        Set<String> keys = redisTemplate.keys(pattern);
-        if (keys != null && !keys.isEmpty()) {
-            redisTemplate.delete(keys);
-        }
     }
 }
