+++ src/main/java/com/giftforyoube/scheduler/Scheduler.java
@@ -3,10 +3,9 @@ package com.giftforyoube.scheduler;
 import com.giftforyoube.funding.entity.Funding;
 import com.giftforyoube.funding.entity.FundingStatus;
 import com.giftforyoube.funding.repository.FundingRepository;
-import com.giftforyoube.funding.service.FundingService;
+import com.giftforyoube.funding.service.CacheService;
 import com.giftforyoube.notification.entity.NotificationType;
 import com.giftforyoube.notification.service.NotificationService;
-import com.giftforyoube.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.cache.annotation.CacheEvict;
@@ -21,13 +20,12 @@ import java.util.List;
 @RequiredArgsConstructor
 public class Scheduler {
 
-    private final FundingService fundingService;
     private final FundingRepository fundingRepository;
     private final NotificationService notificationService;
-    private final UserRepository userRepository;
+    private final CacheService cacheService;
     // 매일 자정에 실행, 마감일이 지난 펀딩의 상태를 업데이트
     // 초, 분, 시, 일, 월, 주 순서
-    @Scheduled(cron = "0 0 0 * * ?")
+    @Scheduled(cron = "0 53 17 * * ?")
     @CacheEvict(value = {"activeMainFundings", "activeFundings", "finishedFundings", "fundingDetail"}, allEntries = true)
     public void autoFinishFundings() {
         log.info("마감일 종료 상태 업데이트 실행");
@@ -45,6 +43,6 @@ public class Scheduler {
             NotificationType notificationType = NotificationType.FUNDING_TIME_OUT;
             notificationService.send(funding.getUser(), notificationType, content, url);
         }
-        fundingService.clearFundingCaches();
+        cacheService.clearFundingCaches();
     }
 }
